Move sqlutil.QuerySignature to apmsql
Expose QuerySignature in the apmsql module.
