+++ module/apmsql/signature.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package sqlutil
+package apmsql
 
 import (
 	"strings"
